refactor: use osrm.Waypoint instead of osrmtype.Waypoint
